requirements.txt
streamlit
langchain
PyPDF2
python-dotenv
faiss-cpu
sentence-transformers
pydantic
boltiotai
